# pkg/trace/writer: pool gzip writers to reduce garbage