<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?
Serialization creates a new gzip writer on every payload, which creates a lot of garbage. The gzip writers can be reset and reused, keeping their internal buffers and dramatically reducing the amount of garbage created during serialization.

This also includes a benchmark, showing this change dramatically reduces the garbage as claimed, and also significantly reduces the cpu time.
```
name                old time/op    new time/op    delta
Serialize/large-16    1.47ms ± 7%    0.99ms ± 5%  -33.16%  (p=0.000 n=9+10)
Serialize/small-16     194µs ±10%      20µs ± 1%  -89.50%  (p=0.000 n=10+10)

name                old alloc/op   new alloc/op   delta
Serialize/large-16    1.64MB ± 0%    0.46MB ± 1%  -71.99%  (p=0.000 n=10+10)
Serialize/small-16    1.21MB ± 0%    0.00MB ± 2%  -99.77%  (p=0.000 n=9+10)

name                old allocs/op  new allocs/op  delta
Serialize/large-16      28.0 ± 0%       8.0 ± 0%  -71.43%  (p=0.000 n=9+10)
Serialize/small-16      27.0 ± 0%       8.0 ± 0%  -70.37%  (p=0.000 n=10+10)
```

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes
This has been run through unit tests and benchmarks.